diary creates diary-* functions, cleaned up diary
The purpose of this is to make any invocation of 'diary command' get trapped
in the original directory that diary was located in. This allows me to have
multiple active versions of my diary scripts without them interacting with
each other. The only downside is that all scripts must now be started with the
'diary command' syntax instead of executing them directly.